Literal return translation for pure functions (#32)
This PR introduces the translation of pure function bodys for any
function that has a literal return. I introduced a new "Return"
embedding, which return statements are now being embedded into.

When converting a pure function's body, the position and block embedding
data is ignored and the most inner embedding is assumed to be a return
embedding. If a non return embedding is found there an error is thrown.
Otherwise, the return expression is being converted to Viper, which is
only successful for literals.