Hook in purity checker (#35)
This PR hooks in the existing purity checking into the translation
workflow of pure functions to validate the pureness of their body.
Namely the util function isPure() was introduced in ExprUtils to run the
purity check on a provided embedding. I introduced a test-case to
validate that there are no false negatives (the passing pure_function
tests validate, that the purity checker has no false positives)

Changes to the existing workflow:

- Non-Pure function bodies, which have impure asserts, now are not being
translated (i. e. an empty body is returned and the translation results
in a header - see test-case)
- For clarity's sake I renamed .isValid() ->
.satisfiesPurityRequirements() and .checkValidity() ->
allNodesSatisfyPurityRequirements() - UPDATE: Dropped this point